apps/backend/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "pymonorepo_proj1" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../coverage/proj1/html' --cov-report xml:'../coverage/proj1/coverage.xml' --html='../reports/proj1/unittests/html/index.html' --junitxml='../reports/proj1/unittests/junit.xml'"

[tool.poetry]
name = "api"
version = "1.0.0"
description = ""
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "backend"

  [tool.poetry.dependencies]
  python = ">=3.9,<3.13"
  supabase = "^2.9.1"
  sqlalchemy = "^2.0.36"
  flask = "^3.0.3"
  flask-restful = "^0.3.10"
  python-dotenv = "^1.0.1"
  psycopg2 = "^2.9.10"
  flask-cors = "^5.0.0"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"
  pytest = "7.3.1"
  pytest-sugar = "0.9.7"
  pytest-cov = "4.1.0"
  pytest-html = "3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
